+++ src/View/Helper/CookieHelper.php
@@ -14,8 +14,6 @@
 
 namespace Origin\View\Helper;
 
-use Origin\Http\Cookie;
-
 /**
  * Cookie Helper - makes it easy to work with cookies, cookies are set using the response
  * object. By default contents of cookies are encrypted.
@@ -23,26 +21,7 @@ use Origin\Http\Cookie;
 
 class CookieHelper extends Helper
 {
-    /**
-     * Cookie Object
-     *
-     * @var \Origin\Http\Cookie
-     */
-    protected $cookie = null;
-
-    /**
-     * Lazy loads the cookie object
-     *
-     * @return \Origin\Http\Cookie
-     */
-    protected function cookie()
-    {
-        if ($this->cookie === null) {
-            $this->cookie = new Cookie();
-        }
-        return $this->cookie;
-    }
-    /**
+  /**
      * Reads a value of a cookie from request
      *
      * @param string $name
@@ -50,7 +29,7 @@ class CookieHelper extends Helper
      */
     public function read(string $name)
     {
-        return $this->cookie()->read($name);
+        return $this->request()->cookies($name);
     }
 
     /**
@@ -78,7 +57,7 @@ class CookieHelper extends Helper
      */
     public function delete(string $name)
     {
-        $this->response()->cookie($name, '','-60 seconds');
+        $this->response()->cookie($name, '', '-60 minutes');
     }
 
     /**
@@ -93,15 +72,16 @@ class CookieHelper extends Helper
         return $this->exists($name);
     }
 
-     /**
-     * Checks if a cookie exists
-     *
-     * @param string $name
-     * @return void
-     */
+    /**
+    * Checks if a cookie exists
+    *
+    * @param string $name
+    * @return void
+    */
     public function exists(string $name) : bool
     {
-        return $this->cookie()->exists($name);
+        $cookies = $this->request()->cookies();
+        return isset($cookies[$name]);
     }
 
     /**